+++ R/createAssoPerm.R
@@ -483,11 +483,11 @@ createAssoPerm <- function(x,
       assoMat1.tmp <- .calcAssociation(count1.tmp,
                                        measure = parNC$measure,
                                        measurePar = parNC$measurePar,
-                                       verbose = FALSE)
+                                       verbose = FALSE)$assoMat
       assoMat2.tmp <- .calcAssociation(count2.tmp,
                                        measure = parNC$measure,
                                        measurePar = parNC$measurePar,
-                                       verbose = FALSE)
+                                       verbose = FALSE)$assoMat
       
       fmat <- fm.open(filenamebase = fileStoreAssoPerm)