+++ tests/testthat/test_netConstruct.R
@@ -26,6 +26,29 @@ plot(testprops)
 mtext("phyloseq", side = 3, cex = 1.5)
 
 #===============================================================================
+context("netConstruct with TreeSummarizedExperiment object as input")
+
+library(mia)
+data("peerj13075")
+
+testnet <- netConstruct(peerj13075,
+                        filtTax = "highestVar",
+                        filtTaxPar = list(highestVar = 20),
+                        filtSamp = "totalReads",
+                        filtSampPar = list(totalReads = 1000),
+                        zeroMethod = "none", normMethod = "none",
+                        measure = "pearson",
+                        sparsMethod = "threshold", thresh = 0.3,
+                        seed = 20190101)
+
+testprops<- netAnalyze(testnet,
+                       clustMethod = "cluster_fast_greedy",
+                       hubPar = "eigenvector")
+
+plot(testprops)
+mtext("phyloseq", side = 3, cex = 1.5)
+
+#===============================================================================
 # Filtering
 
 context("netConstruct with different taxa filtering methods (single network)")
@@ -82,7 +105,8 @@ context("netConstruct with different association measures")
 set.seed(123456)
 
 measures <- c("pearson", "spearman", "bicor", "sparcc", "cclasso", "ccrepe",
-              "propr","gcoda", "spieceasi_gl", "spieceasi_mb", "spring" )
+              "propr","gcoda", "spieceasi_gl", "spieceasi_mb", "spiecease_slr",
+              "spring" )
 
 for (i in 1:length(measures)) {
   
@@ -102,8 +126,14 @@ for (i in 1:length(measures)) {
                        pulsar.params = list(rep.num=5,
                                             thresh = 0.2))
     measure.tmp <- "spieceasi"
+  } else if (measure.tmp == "spiecease_slr") {
+    measurePar <- list(method = "slr",
+                       nlambda=5,
+                       pulsar.params = list(rep.num=5,
+                                            thresh = 0.2))
+    measure.tmp <- "spieceasi"
   } else if (measure.tmp == "spring") {
-    measurePar <- list(nlambda = 5, rep.num = 5)
+    measurePar <- list(nlambda = 5, rep.num = 5, Rmethod = "approx")
   } else if (measure.tmp == "gcoda") {
     measurePar <- list(nlambda = 5)
   } else {
@@ -875,7 +905,7 @@ for (i in 1:length(measures)) {
                                             thresh = 0.3))
     measure.tmp <- "spieceasi"
   } else if (measure.tmp == "spring") {
-    measurePar <- list(nlambda = 5, rep.num = 5)
+    measurePar <- list(nlambda = 5, rep.num = 5, Rmethod = "approx")
   } else if (measure.tmp == "gcoda") {
     measurePar <- list(nlambda = 5)
   } else {